refactor(cli): extract ask/search/chat to cli/_search.py (#284, step 4)
Move the retrieval-facing commands (vstash ask / search / chat, ~635 lines) into
vstash/cli/_search.py; __init__ imports it to register them. __init__ drops the
imports those commands were the last users of (Markdown, chat_module,
search_with_embedding, relevance_tier, _inference_hint, _record_miss_event).

__init__.py: 2216 -> 1581 LOC. Bodies moved verbatim.

- pytest tests/ -> 1290 passed; ruff clean; ask/search/chat resolve via CliRunner
